set Platemap as generic data attribute in kaleido parser
